pyproject.toml
[project]
name = "research-workspace"
version = "0.1.0"
description = "Research workspace for experiments"
requires-python = ">=3.10"
dependencies = [
    "openai",
    "datasets",
    "scikit-learn",
    "numpy",
    "matplotlib",
    "tqdm",
    "pandas",
    "python-dotenv",
    "scipy",
    "transformers", 
    "torch"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
